add circle button with styles
